refactor: Refactor god class - introduce partition and queue state classes

+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/ParallelEoSStreamProcessor.java
@@ -1055,7 +1055,7 @@ public class ParallelEoSStreamProcessor<K, V> implements ParallelStreamProcessor
     }
 
     @Override
-    public int workRemaining() {
+    public long workRemaining() {
         return wm.getNumberOfEntriesInPartitionQueues();
     }
